refactor(sensor): Rename MagEncoder to TalonMagEncoder
Make things clearer
